+++ components/Icons/NuxtStuff.vue
