+++ library/stores/globals.ts
